refactor(dg): create paginationStore
