+++ packages/data-grid/src/stores/index.ts
@@ -1,1 +1,2 @@
 export * from './data'
+export * from './pagination'
